+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- `FileFormats`: File formats get sorted correctly
+
 ## [2.3.4] - 2020-07-16
 
 ### Changed